+++ src/ui-components/basicInput/Text.js
@@ -1,3 +1,4 @@
 import TextBox from './text/TextBox'
 import TextPassword from './text/TextPassword'
-export{TextBox, TextPassword}
+import TextArea from './text/TextArea'
+export{TextBox, TextPassword, TextArea}
