+++ src/ui-components/basicInput/text/TextArea.js
@@ -98,3 +98,69 @@
 		* __Type - __ __[Name]__
 		Description (if necessary)
 **/
+import React, {Component} from 'react'
+import PropTypes from 'prop-types';
+import './Style.css';
+
+
+export default class TextArea extends Component {
+
+    state = {
+        keyword: this.props.value
+    }
+
+    _inputChangedHandler(event) {
+        this.setState({ keyword: event.target.value })
+        this.props.onChangeText(event.target.value)
+    }
+
+    render() {
+        const {
+            rows,
+			cols,
+			style
+		} = this.props
+		const {
+			keyword
+		} = this.state
+        return (
+            <textarea className='textBox'
+                rows={rows}
+                cols={cols}
+				value={keyword}
+				style={style}
+                onChange={(event) => this._inputChangedHandler(event)}
+            >
+            </textarea>
+        )
+    }
+}
+
+TextArea.propTypes = {
+	cols: PropTypes.number,
+	disabled: PropTypes.bool,
+	onChangeText: PropTypes.func,
+	rows: PropTypes.number,
+	style: PropTypes.object,
+	value: PropTypes.string
+};
+
+TextArea.defaultProps = {
+	cols: 50,
+    disabled: false,
+	onChangeText: noAction,
+	rows:5,
+	style: {width: "300px", height: "160px"},
+	value: PropTypes.string
+};
+
+function noAction(mod) {
+    switch (mod) {
+        case 1:
+            console.log('AreaText disabled')
+            break
+        default:
+            console.error('The AreaText State is undefined')
+    }
+
+}
